Move game fixtures to own file
